+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking changes
 
 - Removed `Decoded` struct, inlined its fields into `FieldInfo`.
+- `DecodeError` variants changed.
 
 ### Bugfixes
 
@@ -13,6 +14,7 @@
 ### New features
 
 - Added ISS decoding for SVE exceptions.
+- Added ISS decoding for LD64B/ST64B\* exceptions.
 
 ## 0.1.0